* CC: fix SF ticket 41 CC can`t parse defines with Doxygen single-line comment.
This fixes the doxygen document and token synchronization issue that hasn't
fixed in commit r9905. When reading a macro definition which ends by a C++
comments, we should stop the parsing before the "//" chars, so that the
doxygen document can be added correctly to previous added Macro definition token.
See details in:
https://sourceforge.net/p/codeblocks/tickets/41/

git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@10321 2a5c6006-c6dd-42ca-98ab-0921f2732cef